# openapi.yaml
openapi: 3.0.3
info:
  title: Needle Eye API
  version: 0.1.0
  description: |
    Express + TypeScript backend for the Needle Eye boutique/tailoring ERP.
    Stateless REST: every request carries its own bearer token, verified
    independently -- nothing about a client's session is stored server-side
    between requests. `needleye-web` never talks to Supabase directly; every
    auth operation goes through the `/auth/*` endpoints documented here.

    See the repo README for the layered architecture (Controller ->
    Validation -> Service -> Repository -> Mapper) and the Mermaid flow-map
    diagrams -- this document covers the *what* (request/response contracts),
    the README's Flow Map section covers the *how* (which function calls
    which).
  contact:
    name: Needle Eye API
servers:
  - url: http://localhost:4000/api/v1
    description: Local dev
...
security:
  - bearerAuth: []
paths:
...
  /orders:
    get:
      tags: [Orders]
      summary: List orders (row-scoped by role)
      description: |
        Owner/Manager and Accountant see every order; Designer and Master
        Tailor only ever see orders where they're `designerId`/`masterTailorId`
        -- enforced inside the repository's query construction, not just a
        UI filter.
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/GranularStatus" }
        - name: designerId
          in: query
          schema: { type: string, format: uuid }
        - name: masterTailorId
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Orders visible to the caller
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items: { $ref: "#/components/schemas/Order" }
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      tags: [Orders]
      summary: Create an order
      description: Owner/Manager and Designer only (`orders:create`).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOrderRequest" }
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                type: object
                properties: { order: { $ref: "#/components/schemas/Order" } }
        "403":
          $ref: "#/components/responses/Forbidden"
  /orders/stats:
    get:
      tags: [Orders]
      summary: Dashboard aggregates for the caller's row scope
      description: |
        Registered before `/orders/{id}` -- Express would otherwise match
        `stats` as `:id`. Counts/sums are computed in Postgres (COUNT/SUM
        FILTER), never by fetching every order into Node. `pendingPayments`,
        `collectedRevenue`, and `outstandingRevenue` are absent entirely
        (not zero) for callers without `payments:read` (master_tailor) --
        the same real server-side omission `Order`'s own payment fields get.
      responses:
        "200":
          description: Aggregates scoped to the caller's role (row-scoped the same way GET /orders is)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrderStats" }
        "403":
          $ref: "#/components/responses/Forbidden"
  /orders/{id}:
    get:
      tags: [Orders]
      summary: Get a single order
      description: 404s (not 403) if the order exists but is outside the caller's row scope -- doesn't confirm the order's existence to someone who shouldn't see it.
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: The order
          content:
            application/json:
              schema:
                type: object
                properties: { order: { $ref: "#/components/schemas/Order" } }
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags: [Orders]
      summary: Edit an order
      description: |
        The core field-level RBAC: customer/product fields
        (`orders:edit:customer_product_fields`) and pricing/assignment fields
        (`totalAmount`, `designerId`, `masterTailorId`, gated by
        `orders:edit:pricing_assignment`) are checked *independently* -- a
        Designer can edit their own order's notes but gets 403 touching
        `totalAmount`. `productionStatus` is accepted by the request schema
        but silently stripped by the controller before it reaches the service
        -- status changes go through `PATCH /orders/{id}/status` instead,
        which enforces design-stage vs production-stage RBAC.
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateOrderRequest" }
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                type: object
                properties: { order: { $ref: "#/components/schemas/Order" } }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /orders/{id}/status:
    patch:
      tags: [Orders]
      summary: Change an order's production status
      description: |
        Stage-ownership RBAC, not the flat `orders:edit:*` capabilities:
        moving into a design-stage status (design_pending/design_approved/
        fabric_purchased) requires `orders:status:design_stages`; moving into
        any other (production-stage) status requires
        `orders:status:production_stages`. Deliberately not gated by a
        `requireCapability` middleware call -- which capability applies
        depends on the *target* status in the request body, so the check
        happens entirely inside the service/domain layer
        (assertCanTransitionStatus). Also appends one row to this order's
        status history, in the same DB transaction as the status update.
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateOrderStatusRequest" }
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                type: object
                properties: { order: { $ref: "#/components/schemas/Order" } }
        "403":
          description: Role has no access to this stage at all, or the order isn't assigned to this caller
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          $ref: "#/components/responses/NotFound"
  /orders/{id}/history:
    get:
      tags: [Orders]
      summary: List an order's status-change audit trail
      description: Newest first. Each entry's `label` is a frozen snapshot of the status's label at the time of the change, not recomputed from the current vocabulary.
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Status history entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  history:
                    type: array
                    items: { $ref: "#/components/schemas/OrderStatusHistoryEntry" }
        "404":
          $ref: "#/components/responses/NotFound"
  /orders/{id}/images:
    post:
      tags: [Orders]
      summary: Upload a reference image into one of the 4 slots
      description: Gated by `orders:edit:customer_product_fields` (same permission as editing the order's content fields). `multipart/form-data`, max 10MB, one of 4 slots.
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [slot, file]
              properties:
                slot: { type: integer, minimum: 1, maximum: 4 }
                file: { type: string, format: binary }
      responses:
        "201":
          description: Image stored, returns a fresh signed URL
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrderImage" }
        "400":
          description: Invalid slot, missing file, or unsupported content type
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          $ref: "#/components/responses/Forbidden"
  /orders/{id}/images/{slot}:
    delete:
      tags: [Orders]
      summary: Remove an image from a slot
      parameters:
        - $ref: "#/components/parameters/OrderId"
        - name: slot
          in: path
          required: true
          schema: { type: integer, minimum: 1, maximum: 4 }
      responses:
        "204": { description: Removed }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /orders/{orderId}/payments:
    get:
      tags: [Payments]
      summary: List an order's payment ledger
      description: Gated by `payments:read` at the router level; a scoped (designer) caller additionally must be the order's designer, checked in the service.
      parameters:
        - $ref: "#/components/parameters/OrderIdPath"
      responses:
        "200":
          description: Ledger entries, most recent paid_at first
          content:
            application/json:
              schema:
                type: object
                properties:
                  payments:
                    type: array
                    items: { $ref: "#/components/schemas/Payment" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    post:
      tags: [Payments]
      summary: Record a payment
      description: |
        Gated by `payments:manage`. If the order is currently marked
        `fully_paid`, this is rejected with 409 unless the resulting ledger
        sum would still exactly equal `total_amount` -- see the README's
        Flow Map for the full consistency rule (it runs in both directions:
        here, and on `PATCH /orders/{id}` when *setting* `fully_paid`).
      parameters:
        - $ref: "#/components/parameters/OrderIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePaymentRequest" }
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                type: object
                properties: { payment: { $ref: "#/components/schemas/Payment" } }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Would leave a fully_paid order's ledger not matching its total_amount
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /orders/{orderId}/payments/{paymentId}:
    patch:
      tags: [Payments]
      summary: Edit a payment entry
      description: Same `fully_paid` consistency rule as POST, evaluated against the edited amount.
      parameters:
        - $ref: "#/components/parameters/OrderIdPath"
        - $ref: "#/components/parameters/PaymentId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdatePaymentRequest" }
      responses:
        "200":
          description: Updated payment
          content:
            application/json:
              schema:
                type: object
                properties: { payment: { $ref: "#/components/schemas/Payment" } }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Would leave a fully_paid order's ledger not matching its total_amount
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

    delete:
      tags: [Payments]
      summary: Remove a payment entry
      description: Rejected with 409 if the order is `fully_paid` and removing this entry would break the ledger-sum-equals-total_amount invariant -- change the order's payment status first if that's intentional.
      parameters:
        - $ref: "#/components/parameters/OrderIdPath"
        - $ref: "#/components/parameters/PaymentId"
      responses:
        "204": { description: Removed }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Would leave a fully_paid order's ledger not matching its total_amount
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase-issued access token, obtained from `POST /auth/login` (or `/refresh`, `/exchange-code`). Verified independently on every request -- see `common/middleware/auth.middleware.ts`.
  parameters:
    OrderId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    OrderIdPath:
      name: orderId
      in: path
      required: true
      schema: { type: string, format: uuid }
    PaymentId:
      name: paymentId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Forbidden:
      description: Authenticated, but the caller's role/scope doesn't allow this
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found (or outside the caller's row scope)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      description: The one shape every error response takes, produced by common/middleware/error.middleware.ts regardless of which layer threw.
      properties:
        error: { type: string }
        code: { type: string, example: FORBIDDEN }
        details: {}
...
    GranularStatus:
      type: string
      enum:
        - design_pending
        - design_approved
        - fabric_purchased
        - cutting
        - stitching
        - hand_work
        - machine_work
        - finishing
        - quality_check
        - ready_for_delivery
        - delivered
    ProductCategory:
      type: string
      enum: [designer_blouse, saree, bridal_lehenga, custom_ethnic_wear, boutique_fashion]
    PaymentStatus:
      type: string
      enum: [advance_paid, partially_paid, fully_paid]
    PaymentMethod:
      type: string
      enum: [cash, card, upi, bank_transfer, cheque, other]
...
    Payment:
      type: object
      properties:
        id: { type: string, format: uuid }
        orderId: { type: string, format: uuid }
        amount: { type: number, description: Always > 0 -- a correction is a payment update/delete, not a negative entry. }
        method: { $ref: "#/components/schemas/PaymentMethod" }
        paidAt: { type: string, format: date }
        recordedBy: { type: string, format: uuid, nullable: true }
        recordedByName: { type: string }
        notes: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    CreatePaymentRequest:
      type: object
      required: [amount, method]
      properties:
        amount: { type: number, minimum: 0, exclusiveMinimum: true }
        method: { $ref: "#/components/schemas/PaymentMethod" }
        paidAt: { type: string, format: date, description: Defaults to today if omitted. }
        notes: { type: string }
    UpdatePaymentRequest:
      type: object
      description: Every field from CreatePaymentRequest, but optional (partial update).
      properties:
        amount: { type: number, minimum: 0, exclusiveMinimum: true }
        method: { $ref: "#/components/schemas/PaymentMethod" }
        paidAt: { type: string, format: date }
        notes: { type: string }
    UpdateOrderStatusRequest:
      type: object
      required: [status]
      properties:
        status: { $ref: "#/components/schemas/GranularStatus" }
    OrderStatusHistoryEntry:
      type: object
      properties:
        id: { type: string, format: uuid }
        orderId: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/GranularStatus" }
        label: { type: string, description: Snapshot of the status's label at the time of the change. }
        changedBy: { type: string, format: uuid, nullable: true }
        changedByName: { type: string }
        createdAt: { type: string, format: date-time }
    OrderImage:
      type: object
      properties:
        id: { type: string, format: uuid }
        orderId: { type: string, format: uuid }
        slot: { type: integer, minimum: 1, maximum: 4 }
        storagePath: { type: string }
        url: { type: string, description: Freshly-signed Supabase Storage URL, re-resolved on every read -- never stored. }
        originalFilename: { type: string, nullable: true }
        contentType: { type: string, nullable: true }
        sizeBytes: { type: integer, nullable: true }
        uploadedBy: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }
    Order:
      type: object
      properties:
        id: { type: string, format: uuid }
        orderNumber: { type: string, example: ORD-2026-014 }
        customerName: { type: string }
        phone: { type: string }
        billNumber: { type: string }
        bookingDate: { type: string, format: date }
        dueDate: { type: string, format: date }
        designerId: { type: string, format: uuid }
        designerName: { type: string }
        masterTailorId: { type: string, format: uuid }
        masterTailorName: { type: string }
        productCategory: { $ref: "#/components/schemas/ProductCategory" }
        orderDetails: { type: string }
        handWork: { type: boolean }
        machineWork: { type: boolean }
        purchaseRequired: { type: boolean }
        paymentStatus:
          allOf: [{ $ref: "#/components/schemas/PaymentStatus" }]
          description: Absent entirely (not null) for callers without payments:read (master_tailor) -- a real server-side omission, not a UI-only hide.
        totalAmount: { type: number, description: "Absent for master_tailor -- see paymentStatus." }
        amountPaid: { type: number, description: "Absent for master_tailor. The real sum of GET /orders/{orderId}/payments -- never stored, always derived at read time." }
        outstanding: { type: number, description: "Absent for master_tailor -- see paymentStatus." }
        productionStatus: { $ref: "#/components/schemas/GranularStatus" }
        designerInstructions: { type: string, nullable: true }
        specialNotes: { type: string, nullable: true }
        images:
          type: array
          items: { $ref: "#/components/schemas/OrderImage" }
        createdBy: { type: string, format: uuid, nullable: true }
        updatedBy: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    OrderStats:
      type: object
      properties:
        total: { type: integer, description: Every order in the caller's row scope. }
        active: { type: integer, description: "Not yet in a completed canonical stage (ready_for_delivery/delivered)." }
        completed: { type: integer, description: In the ready_for_delivery or delivered canonical stage. }
        pendingPayments: { type: integer, description: "Absent for master_tailor. Orders whose paymentStatus isn't fully_paid." }
        collectedRevenue: { type: number, description: "Absent for master_tailor. SUM(payments.amount) across the scope." }
        outstandingRevenue: { type: number, description: "Absent for master_tailor. SUM(orders.total_amount) - collectedRevenue, floored at 0." }
    CreateOrderRequest:
      type: object
      required:
        - customerName
        - phone
        - billNumber
        - dueDate
        - designerId
        - masterTailorId
        - productCategory
        - orderDetails
        - paymentStatus
        - productionStatus
      properties:
        customerName: { type: string }
        phone: { type: string, pattern: "^\\d{10}$" }
        billNumber: { type: string }
        bookingDate: { type: string, format: date }
        dueDate: { type: string, format: date }
        designerId: { type: string, format: uuid }
        masterTailorId: { type: string, format: uuid }
        productCategory: { $ref: "#/components/schemas/ProductCategory" }
        orderDetails: { type: string }
        handWork: { type: boolean, default: false }
        machineWork: { type: boolean, default: false }
        purchaseRequired: { type: boolean, default: false }
        paymentStatus: { $ref: "#/components/schemas/PaymentStatus" }
        totalAmount: { type: number, minimum: 0, default: 0 }
        productionStatus: { $ref: "#/components/schemas/GranularStatus" }
        designerInstructions: { type: string }
        specialNotes: { type: string }
    UpdateOrderRequest:
      type: object
      description: |
        Every field from CreateOrderRequest, but optional (partial update) --
        deliberately a single role-agnostic schema rather than one per role.
        Which fields a given caller is actually allowed to touch is enforced
        by the service layer's RBAC field-splitting (see PATCH /orders/{id}),
        not by this schema. `productionStatus` is accepted here but the
        controller strips it before the service ever sees it.
      properties:
        customerName: { type: string }
        phone: { type: string, pattern: "^\\d{10}$" }
        billNumber: { type: string }
        bookingDate: { type: string, format: date }
        dueDate: { type: string, format: date }
        designerId: { type: string, format: uuid }
        masterTailorId: { type: string, format: uuid }
        productCategory: { $ref: "#/components/schemas/ProductCategory" }
        orderDetails: { type: string }
        handWork: { type: boolean }
        machineWork: { type: boolean }
        purchaseRequired: { type: boolean }
        paymentStatus: { $ref: "#/components/schemas/PaymentStatus" }
        totalAmount: { type: number, minimum: 0 }
        productionStatus: { $ref: "#/components/schemas/GranularStatus" }
        designerInstructions: { type: string }
        specialNotes: { type: string }